Move CPakFile to own implementation file
* Moved CPakFile class to own implementation file.
* Moved common constants to public header.
