+++ src/public/const.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#define DEFAULT_RPAK_NAME "new"
+#define PF_KEEP_DEV 1 << 0 // whether or not to keep debugging information
